# rebuild: refactor index.js using FP patterns - significant improvements in package size and performance