--- a/src/index.js
+++ b/src/index.js
@@ -1,40 +1,30 @@
-function toVal(mix) {
-	var k, y, str='';
-
-	if (typeof mix === 'string' || typeof mix === 'number') {
-		str += mix;
-	} else if (typeof mix === 'object') {
-		if (Array.isArray(mix)) {
-			for (k=0; k < mix.length; k++) {
-				if (mix[k]) {
-					if (y = toVal(mix[k])) {
-						str && (str += ' ');
-						str += y;
-					}
-				}
+function toStr(mix) {
+	switch (mix && typeof mix) {
+		case 'number':
+		case 'string':
+			return '' + mix;
+		case 'object':
+			if (Array.isArray(mix)) {
+				return digest(mix);
 			}
-		} else {
-			for (k in mix) {
-				if (mix[k]) {
-					str && (str += ' ');
-					str += k;
-				}
-			}
-		}
+		return Object.keys(mix)
+			.filter(function(key) {
+				return mix[key];
+			})
+			.join(' ');
+		default:
+			return '';
 	}
-
-	return str;
 }
-
-export default function () {
-	var i=0, tmp, x, str='';
-	while (i < arguments.length) {
-		if (tmp = arguments[i++]) {
-			if (x = toVal(tmp)) {
-				str && (str += ' ');
-				str += x
-			}
-		}
-	}
-	return str;
+  
+function digest(arr) {
+	return arr.map(toStr)
+		.filter(function(str) {
+			return str;
+		})
+	.join(' ');
+}
+  
+export default function() {
+	return digest(Array.prototype.slice.call(arguments));
 }