@lukeed 2021-07-06
Hi, thank you, but your changes are significantly slower. They seem to be the last line per bench, and higher "ops/sec" is better.

Functional primitives are (almost) aways slower than for loops in JS.