@malowitzki 2021-07-06
ah interesting, I thought ops/sec was a representation of load, so was shooting for lowest possible load — got that backwards